+++ security/account/admin.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+admin_users=( $(sudo groupmems --group wheel --list) )
+
+echo wheel: ${admin_users[@]}