+++ security/account/password.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+# sudo dnf install -y libpwquality
+
+random_pw=(
+	$(pwmake 64)
+	$(pwmake 128)
+	$(pwmake 256)
+)
+
+echo random password: ${random_pw[@]}